+++ packages/apps/intervention/src/components/atoms/EquipmentCard/EquipmentCard.tsx
