+++ packages/apps/intervention/src/components/atoms/index.ts
@@ -16,6 +16,5 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export {default as ClientProspectSearchBar} from './ClientProspectSearchBar/ClientProspectSearchBar';
+export {default as EquipmentCard} from './EquipmentCard/EquipmentCard';
 export {default as InterventionCard} from './InterventionCard/InterventionCard';
-export {default as PlaceEquipmentSearchBar} from './PlaceEquipmentSearchBar/PlaceEquipmentSearchBar';
