+++ packages/apps/intervention/src/components/molecules/index.ts
@@ -16,5 +16,5 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export {default as EquipmentCard} from './EquipmentCard/EquipmentCard';
+export {default as EquipmentActionCard} from './EquipmentActionCard/EquipmentActionCard';
 export {default as InterventionDetailCard} from './InterventionDetailCard/InterventionDetailCard';
