+++ packages/apps/intervention/src/components/organisms/ClientProspectSearchBar/ClientProspectSearchBar.js
