+++ packages/apps/intervention/src/components/organisms/PlaceEquipmentSearchBar/PlaceEquipmentSearchBar.js
