+++ packages/apps/intervention/src/components/organisms/index.js
@@ -16,4 +16,5 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export {default as EquipmentActionCard} from './EquipmentActionCard/EquipmentActionCard';
+export {default as ClientProspectSearchBar} from './ClientProspectSearchBar/ClientProspectSearchBar';
+export {default as PlaceEquipmentSearchBar} from './PlaceEquipmentSearchBar/PlaceEquipmentSearchBar';
